@@ -29,62 +29,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.13",
-    "@angular/compiler": "^7.2.13",
-    "@angular/core": "^7.2.13",
-    "@angular/forms": "^7.2.13",
-    "@angular/http": "^7.2.13",
-    "@angular/platform-browser": "^7.2.13",
-    "@angular/platform-browser-dynamic": "^7.2.13",
-    "@angular/platform-server": "^7.2.13",
-    "@angular/router": "^7.2.13",
-    "ag-grid-angular": "^20.1.0",
-    "ag-grid-community": "^20.1.0",
+    "@angular/animations": "^8.2.11",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.11",
+    "@angular/compiler": "^8.2.11",
+    "@angular/core": "^8.2.11",
+    "@angular/forms": "^8.2.11",
+    "@angular/platform-browser": "^8.2.11",
+    "@angular/platform-browser-dynamic": "^8.2.11",
+    "@angular/platform-server": "^8.2.11",
+    "@angular/router": "^8.2.11",
+    "ag-grid-angular": "^21.2.2",
+    "ag-grid-community": "^21.2.2",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.6.5",
-    "highcharts": "^7.1.1",
+    "core-js": "^3.3.2",
+    "highcharts": "^7.2.0",
     "highcharts-angular": "^2.4.0",
     "highcharts-export-csv": "^1.4.3",
     "ion-rangeslider": "^2.3.0",
-    "ng-recaptcha": "^4.2.1",
-    "popper.js": "^1.14.7",
-    "primeicons": "^1.0.0",
-    "primeng": "^7.0.5",
-    "rxjs": "^6.4.0",
-    "tether": "^1.4.5",
+    "ng-recaptcha": "^5.0.0",
+    "popper.js": "^1.16.0",
+    "primeicons": "^2.0.0",
+    "primeng": "^8.0.4",
+    "rxjs": "^6.5.3",
+    "tether": "^1.4.7",
     "tlite": "^0.1.9",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.13.2",
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular-devkit/core": "7.3.2",
-    "@angular-devkit/schematics": "7.3.2",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.13",
-    "@types/google.analytics": "0.0.39",
-    "@types/highcharts": "^5.0.36",
-    "@types/jasmine": "~3.3.9",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/jquery": "^3.3.29",
-    "@types/node": "~11.9.4",
-    "ajv": "^6.9.1",
-    "codelyzer": "~4.5.0",
-    "hoek": "6.1.2",
-    "jasmine-core": "~3.3.0",
+    "@angular-devkit/architect": "0.802.1",
+    "@angular-devkit/build-angular": "^0.802.2",
+    "@angular-devkit/core": "8.2.1",
+    "@angular-devkit/schematics": "8.2.1",
+    "@angular/cli": "^8.3.12",
+    "@angular/compiler-cli": "^8.2.11",
+    "@types/google.analytics": "0.0.40",
+    "@types/highcharts": "^5.0.44",
+    "@types/jasmine": "^3.4.4",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/jquery": "^3.3.31",
+    "@types/node": "^12.7.12",
+    "ajv": "^6.10.2",
+    "codelyzer": "^5.1.2",
+    "hoek": "6.1.3",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "^4.4.0",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.2",
-    "ts-node": "~8.0.2",
-    "tslint": "~5.12.1",
-    "typescript": "3.2.4"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "^3.5.3"
   }
 }
